fixes code climate issues
